requirements.txt
matplotlib==3.5.1
numpy==1.22.3
tensorflow==2.8.0
scikit-learn==1.0.2
Pillow==9.1.0
pandas==1.4.2
scipy==1.8.0
keras==2.8.0
